Major Changes
Refactors back to JS module and got the imports/exports working.
Creates the HTML list to display that added Task items.
Added JS to append the new Tasks to the list with the relevant text content form the object.
Added a checkbox to the new Tasks HTML <li> element set an onclick to toggle completed.
Working on the logic to update the associated Task object's isComplete value.
Tried to add relevant comments to the existing code and improve readability.