Major Changes
Adds a save button and functionality using JSON.
Fixes issues with the content not loading correctly and displaying all text content as "undefined".
Adds a new render function render all the saved task items as well as re-render the list when new tasks are added.
Adds relevant classes to the elements created by the render function to be styled.